fix clippy warnings from switching to Rust 1.52.1
From is now prefered over Into because because implementing From
provides the Into implementation as well by default.

Explicitly define the type of constants when initializing them instead
of using `as` to convert them to the desired type.

+++ src/mmap.rs
@@ -812,10 +812,7 @@ mod tests {
 
     #[test]
     fn test_overlapping_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(0), 100 as usize),
-            (GuestAddress(99), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(0), 100_usize), (GuestAddress(99), 100_usize)];
 
         assert_eq!(
             format!(
@@ -858,10 +855,7 @@ mod tests {
 
     #[test]
     fn test_unsorted_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(100), 100 as usize),
-            (GuestAddress(0), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(100), 100_usize), (GuestAddress(0), 100_usize)];
 
         assert_eq!(
             format!(
@@ -904,10 +898,7 @@ mod tests {
 
     #[test]
     fn test_valid_memory_regions() {
-        let regions_summary = [
-            (GuestAddress(0), 100 as usize),
-            (GuestAddress(100), 100 as usize),
-        ];
+        let regions_summary = [(GuestAddress(0), 100_usize), (GuestAddress(100), 100_usize)];
 
         let guest_mem = GuestMemoryMmap::new();
         assert_eq!(guest_mem.regions.len(), 0);
